moved authentication interfaces (contexts) to a new file
